Fix other plugins loading
It was noticed that file_line plugin breaks the loading of
other plugins. For example, when having next line in ~/.vimrc:

    au BufNewFile main.c silent! 0r ~/.vim/skeleton/template.c

this line loads the content from template file when I'm creating new
main.c file. But when file-line plugin is installed this functionality
doesn't work (newly created main.c is blank). This patch fixes that.

Also, it was reported that file_line breaks vim-go plugin. I haven't
check if this patch fixes it though.

Fixes: #62

Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>